requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib
python-jose
pydantic
python-dotenv
